CXP-1011: Move Install
